refactor: removed mouse factory 2
