--- Assets/Scripts/Mouse/MouseFactory2.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 5d938b9964d248e4e9acf897806b3d1c
-timeCreated: 1723925564
